Add files via upload
This is the main file where vgg16 is getting trained, validated and predicting into 3 classes Plantation, NotPlantation and TreeCover. Then the images with labels which are being predicted as Plantation are further predicted using YOLOv8s model to identify each plant in the subimage and classify it as Dead, good or stunted. This also gives counts for each class and total plantation cost in each grid image as well.